modules/locations_service/requirements.txt
GeoAlchemy2==0.8.4
greenlet==2.0.2
grpcio==1.56.0
kafka-python==2.0.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.23.1
packaging==23.1
protobuf==4.23.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Shapely==1.7.0
SQLAlchemy==1.3.19
